fix: replace std::map containing the Modbus configuration with a flat list
This is a memory optimization. Instead of the map use a flat list and use a binary search algorithm to find the mapping modbus register -> entity short name. This requires the list to be sorted. Added a check before starting the Modbus component that the entries are in the correct order.